validate sample rate, rewrite shouldSample method
Previously, if you had a sample rate of 0.01, then
Math.round(1 / 0.01) = 100 and shouldSample method was doing

ThreadLocalRandom.current().nextInt(samplingRate) == 0

which picked a number from 0 (inclusive) to 100 (exclusive).

However, if we set sample rate to, for example, 0.75 to say that
75% should be sampled, then Math.round(1 / 0.75) = 1.33 and rounded it is 1.

ThreadLocalRandom.current().nextInt(1) == 0

will be always true. That means what basically from some sample rate which rounds to 1
we lose the probability.

The current approach works with floats and it is rewritten to

ThreadLocalRandom.current().nextFloat() < samplingRate

nextFloat() gives values between zero (inclusive) and one (exclusive).
If we set sampling rate to 1 then it will be always true. If we set it to 0.01 that will be 1%.
If we set it to 0.75 that will be 75%, without losing any accuracy.